requirements.txt
alembic==1.10.3
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
orjson==3.8.10
pandas==2.0.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing-extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
